client/pyproject.toml
[tool.poetry]
name = "blindbox"
version = "0.1.0"
description = ""
authors = ["Lucas Bourtoule <lucas.bourtoule@mithrilsecurity.io>"]
readme = "README.md"

[tool.poetry.dependencies]
cbor2 = { version = "^5.4.6", optional = true }
cryptography = { version = "^39.0.2", optional = true }
pycose = { version = "^1.0.1", optional = true }
pydantic = { version = "^1.10.7", optional = true }
pyOpenSSL = { version = "^23.1.1", optional = true }
python = "^3.9"
requests = { version = "^2.28.2", optional = true }
pyyaml = { version = "^6.0" }

# certifi==2022.12.7
# charset-normalizer==3.1.0
# docker==6.0.1
# idna==3.4
# packaging==23.1
# requests==2.28.2
# urllib3==1.26.15
# websocket-client==1.5.1
[tool.poetry.extras]
ai = ["cbor2", "cryptography", "pycose", "pydantic", "pyOpenSSL", "requests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
blindbox = 'blindbox.command.builder:main'
